// admin-frontend/src/api/javaedition/model/onlineUserVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */

/**
 * 在线用户信息
 */
export interface OnlineUserVO {
  /** 用户ID */
  userId?: number;
  /** 用户名 */
  username?: string;
  /** 昵称 */
  nickname?: string;
  /** 头像URL */
  avatar?: string;
  /** 登录IP */
  loginIp?: string;
  /** User-Agent */
  userAgent?: string;
  /** 登录时间 */
  loginTime?: string;
  /** 最后活跃时间 */
  lastActiveTime?: string;
}